Create MathSciStudentCount_LgScale.sql
Created queries that resemble the MathSci Student Count query but with a larger and existing database on my computer (StackOverFlow2013). 

WHERE and HAVING were explored as they impact query performance.

Next we will explore how to avoid using LEFT or other formatting functions in the current query and testing the performance in StackOverFlow.